% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_generator.R
\name{ag_generator}
\alias{ag_generator}
\title{generate an Attack Graph}
\usage{
ag_generator(
  filename.str = system.file("extdata", "example.data", package = "attackgraph"),
  numThreads = 1,
  initQSize = 1,
  so.str = system.file("lib", "app.so", package = "attackgraph")
)
}
\arguments{
\item{filename.str}{The name of the file that describes a target network. Default an example file describing a network with 4 vulnerable hosts.}

\item{numThreads}{The number of threads that accelerate Attack Graph generation. Default 1}

\item{initQSize}{The number of initial nodes for multiple threads to expand, and must be no less than the number of threads. Default 1}

\item{so.str}{The name of the shared object (.so) file for accelerating Attack Graph generation. Default "app.so"}
}
\value{
A list of node matrix, edge matrix, number of nodes and number of edges
}
\description{
This function loads a file describing a target network as input,
and generates its Attack Graph with nodes, edges, number of nodes
and number of edges reported.
}
\details{
The target network described by the file filename.str should follow a predefined format. Invoke format_ref()
function to refer to an example
}
\examples{
ag_generator() # all arguments on default to generate AG for a built-in example target network

ag_generator(create_network(5), 1, 1) # create a target network with 5 vulnerable hosts and generate the AG
with 1 thread and 1 initial node.

ag_generator("user_network.data", n, m) # generate AG based on user defined network data, n threads and
m initial nodes are set. Note n<=m
}
