% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_generator.R
\name{create_network}
\alias{create_network}
\title{create a data file for a target network.}
\usage{
create_network(nv = 5)
}
\arguments{
\item{nv}{The number of vulnerable hosts, default 5}
}
\value{
The path name of the generated data file
}
\description{
This function creates a data file for a target network. The target network follows a fixed pattern
to grow in size, and the number of vulnerable hosts (attacker not included) can be set through an argument.
}
\details{
This function creates target networks described by their data files for testing the attack graph
generator. The data file can be directly read by the ag_generator() function. Pass the returned path name to
the first argument of the ag_generator() function and the corresponding AG can be generated. Note that the
number of vulnerable hosts determines the execution time. While it is safe to set a small nv between 3 and
100, larger values are not recommended.
}
\examples{
filename <- create_network(5) # create the data file for a target network with 5 vulnerable hosts
}
