% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag2ag.R
\name{dag2ag}
\alias{dag2ag}
\title{create a target network for the ag_generator() function based on user-defined directed acyclic graph (DAG)}
\usage{
dag2ag(dag.filename, nw.filename = "user_nw.data")
}
\arguments{
\item{dag.filename}{The file name of the user DAG, which should provide an edge list}

\item{nw.filename}{The file name for the target network. Default "user_nw.data".}
}
\description{
This function creates a target network from a user-defined directed acyclic graph and the data file of the
target network can be used as input to the ag_generator() function.
}
\details{
This function takes a DAG as input, labels its edges with asset/exploit ids. If the labeling process
is successful, the labeled DAG is then converted to a target network described in a data file, which can be
used as the input of the ag_generator() function. If the labeling process fails, no data file will be created.
Note that the adjacency matrix of the input DAG should be an upper triangle matrix with all its diagonal
entry being 0. Moreover, there must be a path from node 1 to all other nodes.
}
\examples{
dag2ag("mydag.txt", "my_nw.data") # read a DAG's edge list from "mydag.txt" and convert it to a
target network described by the data file "my_nw.data"
}
