% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_analysis.R
\name{EKPC}
\alias{EKPC}
\title{Analyze an Attack Graph data with edge kpath entrality (EKPC)}
\usage{
EKPC(k.value, edge.matrix)
}
\arguments{
\item{k.value}{The length k of the paths being counted for each edge}

\item{edge.matrix}{The Matrix of Attack Graph edges provided by the output of ag_generator() function.}
}
\value{
A list of exploit EKPC values, asset EKPC values, edge EKPC values
}
\description{
This function computes edge kpath centrality (EKPC) for given Attack Graphs
}
\details{
This function computes the edge kpath centrality for all the edges in a given Attack Graph. The total
kpath for each exploit/asset is also computed by accumulating the kpath of edges labeled with the same
exploit/asset id.
}
\examples{
EKPC(2, user_attack_graph$edge.matrix) # compute 2-path values for a user Attack Graph
}
