% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_generator.R
\name{format_ref}
\alias{format_ref}
\title{display an example that describes a target network}
\usage{
format_ref()
}
\description{
This function loads a built-in data file from the package and describes an example target network.
Please follow its format to describe your own target network and feed the data file as the input to
the ag_generator() function
}
\details{
The example target network is described by 6 components. The list of assets describes all the
devices in the network. The list of qualities describes the configuration information of each device. The
list of topologies describes the interconnections between assets. The list of exploits describes each
vulnerability name and its class (either local/1 or remote/2). The list of exploit preconditions describes
the requirements of each vulnerability to be successfully exploited. The list of exploit postconditions
describes the consequences of each vulnerablity exploitation.
}
\examples{
format_ref() # display the content of the data file
}
